+++ src/main/java/power/keepeersofthestones/procedures/JoinPlayerInTheWorldProcedure.java
@@ -4,7 +4,7 @@ import power.keepeersofthestones.world.inventory.ChoiseMagicStoneGUIMenu;
 import power.keepeersofthestones.network.PowerModVariables;
 import power.keepeersofthestones.init.PowerModGameRules;
 
-import net.minecraftforge.network.NetworkHooks;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.eventbus.api.Event;
@@ -45,7 +45,7 @@ public class JoinPlayerInTheWorldProcedure {
 					.orElse(new PowerModVariables.PlayerVariables())).selected) {
 				{
 					if (entity instanceof ServerPlayer _ent) {
-						BlockPos _bpos = new BlockPos(x, y, z);
+						BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
 						NetworkHooks.openGui((ServerPlayer) _ent, new MenuProvider() {
 							@Override
 							public Component getDisplayName() {
