+++ src/main/java/power/keepeersofthestones/procedures/UpdateToLevel2Procedure.java
@@ -3,7 +3,7 @@ package power.keepeersofthestones.procedures;
 import power.keepeersofthestones.world.inventory.LevelsAndSkillsPageMenu;
 import power.keepeersofthestones.network.PowerModVariables;
 
-import net.minecraftforge.network.NetworkHooks;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
 
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.item.ItemStack;
@@ -66,7 +66,7 @@ public class UpdateToLevel2Procedure {
 			}
 			{
 				if (entity instanceof ServerPlayer _ent) {
-					BlockPos _bpos = new BlockPos(x, y, z);
+					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
 					NetworkHooks.openGui((ServerPlayer) _ent, new MenuProvider() {
 						@Override
 						public Component getDisplayName() {
